Refatoração: Implementação da Momentus Store (Marketplace), migração para Long XP e melhorias na navegação
Este commit introduz o sistema de loja ao aplicativo, permitindo que usuários troquem pontos acumulados por itens colecionáveis (adesivos e medalhas). Além disso, realiza uma migração técnica do sistema de pontuação para o tipo `Long` para garantir escalabilidade e refina a estrutura de navegação e identidade visual de componentes globais.

**Principais alterações:**

*   **Momentus Store (Marketplace)**:
    *   **`MarketScreen.kt` & `MarketViewModel.kt`**: Nova interface de loja com visualização em grid, exibindo itens categorizados por raridade (Comum, Raro, Épico, Lendário) e status de posse.
    *   **`MarketRepository.kt`**: Gerenciamento de inventário via Firestore. Implementa a lógica de transação: verificação de saldo, dedução de pontos e salvamento no inventário do usuário.
    *   **`MarketItem.kt`**: Definição de modelos de dados para colecionáveis, incluindo multiplicadores de preço baseados em raridade.

*   **Migração de Pontuação (XP/Points)**:
    *   **Tipo de Dado**: Alteração em toda a pilha do app (de `Int` para `Long`) para os campos de pontos e XP, abrangendo `UserData.kt`, `Achievement.kt`, `UserLevel.kt` e seus respectivos repositórios e ViewModels.
    *   **`GamificationRepository.kt`**: Atualização dos métodos de desbloqueio de conquistas para suportar a nova tipagem de incremento de valores no Firestore.

*   **Navegação e Interface (UI)**:
    *   **`AppNavigation.kt`**:
        *   Inclusão da nova rota para a `MarketScreen`.
        *   Refatoração do `Scaffold` para corrigir o preenchimento (padding) do background, eliminando faixas visuais indesejadas.
        *   Estilização da `NavigationBar` com cores mais suaves e indicadores de seleção refinados.
    *   **`MoreScreen.kt`**: Adicionado atalho para acesso à loja na lista de opções.
    *   **`SettingsScreen.kt`**: Ajuste no código hexadecimal da cor "Verde" para consistência com a paleta `EmeraldNeon`.

*   **Infraestrutura e Dados**:
    *   **`UserRepository.kt`**: Adicionado o método `getUserDataSync()` para recuperação imediata de dados do usuário, essencial para validações de compra em tempo real.
    *   **`Color.kt` & `Theme.kt`**: Ajustes finos na paleta "Emerald" e nas propriedades de cor do tema claro/escuro para melhorar o contraste de ícones sobre botões primários.

*   **Manutenção**:
    *   Padronização de animações de conclusão e lógica de observação de pontos no `CalendarViewModel.kt`.
    *   Atualização da lista de conquistas estáticas em `AchievementsList.kt` para a nova tipagem `Long`.

+++ app/src/main/java/br/com/fabriciolima/momentus/data/model/Achievement.kt
@@ -11,5 +11,5 @@ data class Achievement(
     val id: String,
     val name: String,
     val description: String,
-    val points: Int
+    val points: Long
 )
