+++ app/src/main/java/br/com/fabriciolima/momentus/domain/model/UserLevel.kt
@@ -8,27 +8,27 @@ import androidx.compose.ui.graphics.Color
 data class UserLevel(
     val level: Int,
     val rankName: String,
-    val currentXp: Int,
-    val nextLevelXp: Int,
+    val currentXp: Long,
+    val nextLevelXp: Long,
     val progress: Float,
     val rankColor: Color
 ) {
     companion object {
-        fun fromPoints(points: Int): UserLevel {
+        fun fromPoints(points: Long): UserLevel {
             return when {
-                points < 100 -> UserLevel(1, "Iniciante", points, 100, points / 100f, Color(0xFF9E9E9E))
-                points < 300 -> UserLevel(2, "Determinado", points - 100, 200, (points - 100) / 200f, Color(0xFF4CAF50))
-                points < 600 -> UserLevel(3, "Focado", points - 300, 300, (points - 300) / 300f, Color(0xFF2196F3))
-                points < 1000 -> UserLevel(4, "Constante", points - 600, 400, (points - 600) / 400f, Color(0xFF9C27B0))
+                points < 100 -> UserLevel(1, "Iniciante", points, 100L, points / 100f, Color(0xFF9E9E9E))
+                points < 300 -> UserLevel(2, "Determinado", points - 100, 200L, (points - 100) / 200f, Color(0xFF4CAF50))
+                points < 600 -> UserLevel(3, "Focado", points - 300, 300L, (points - 300) / 300f, Color(0xFF2196F3))
+                points < 1000 -> UserLevel(4, "Constante", points - 600, 400L, (points - 600) / 400f, Color(0xFF9C27B0))
                 else -> {
                     val excessPoints = points - 1000
-                    val levelsAbove5 = excessPoints / 500
+                    val levelsAbove5 = (excessPoints / 500).toInt()
                     val xpInCurrentLevel = excessPoints % 500
                     UserLevel(
                         level = 5 + levelsAbove5,
                         rankName = "Mestre da Rotina",
                         currentXp = xpInCurrentLevel,
-                        nextLevelXp = 500,
+                        nextLevelXp = 500L,
                         progress = xpInCurrentLevel / 500f,
                         rankColor = Color(0xFFFFD700)
                     )
@@ -39,7 +39,7 @@ data class UserLevel(
         /**
          * Verifica se houve mudança de nível entre dois valores de pontuação.
          */
-        fun didLevelUp(oldPoints: Int, newPoints: Int): Boolean {
+        fun didLevelUp(oldPoints: Long, newPoints: Long): Boolean {
             if (newPoints <= oldPoints) return false
             return fromPoints(newPoints).level > fromPoints(oldPoints).level
         }
